+++ NEWS.md
@@ -1,6 +1,7 @@
+# enrichplot 1.17.2
 
-# enrichplot 1.17.1.004
-
++ `autofacet` to automatically split `barplot` and `dotplot` into several facets (2022-09-06, Tue)
++ `dotplot` method for `enrichResultList` object 
 + add parameters `hilight_category`, `alpha_hilight`, `alpha_nohilight` for `cnetplot()` and `emapplot` (2022-09-4, Sun)
 + change round digits of cnetplot scatterpie legend to 1 (2022_8_29, Mon).
 + `gsearank()` can export result as a table when `output = "table"` (2022-08-29, Mon, #184)